chore(surveys): converge bar-color logic and address review nits
- route both parents' bar colors through computeBarColors (single source;
  removes the duplicated dim logic and bare 0.22/0.35 alphas)
- MultipleChoiceBarChart now takes barColors as a prop instead of recomputing
- extract formatCountWithPercentage, shared across all four chart formatters
- tighten RatingBarChart data prop to the real ChoiceQuestionResponseData[]
- trim redundant story comments and the computeBarColors JSDoc
